split buildkit client code out into files more
